+++ src/main/java/pl/okazje/project/entities/Tag.java
@@ -8,7 +8,7 @@ public class Tag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer tag_id;
+    private Long tag_id;
     private String name;
 
     @OneToMany(mappedBy="tag")
@@ -26,11 +26,11 @@ public class Tag {
 
     }
 
-    public Integer getTag_id() {
+    public Long getTag_id() {
         return tag_id;
     }
 
-    public void setTag_id(Integer tag_id) {
+    public void setTag_id(Long tag_id) {
         this.tag_id = tag_id;
     }
